Wisconsin Diagnostic Breast Cancer (WDBC).

A dataset containing diagnosis (Malign/benign) for 569 patients with breast cancer and features extracted from tumor imaging.

Usage

wdbc.data

Format

A data frame with 569 rows and 12 variables:

person_id

ID number.

diagnosis

M = malignant, B = benign

radius

Mean of distances from center to points on the perimeter

texture

Standard deviation of gray-scale values.

perimeter

Computed perimeter

area

see doc

smoothness

local variation in radius lengths

compactness

perimeter^2 / area - 1.0

concavity

Severity of concave portions of the contour

concave_points

Number of concave portions of the contour

symmetry

see doc

fractal_dimension

"coastline approximation" - 1

compactness_binary

Factor containing binarized compactness (threshold=0.1). Additional toy parameter.

compactness_quartile

Factor containing compactness divided in quartiles. Additional toy parameter.

Source

https://archive.ics.uci.edu/ml/datasets/Breast+Cancer+Wisconsin+(Diagnostic)

Details

Creators:

Dr. William H. Wolberg

General Surgery Dept., University of Wisconsin, Clinical Sciences Center, Madison, WI 53792

W. Nick Street

Computer Sciences Dept., University of Wisconsin, 1210 West Dayton St., Madison, WI 53706

Olvi L. Mangasarian

Computer Sciences Dept., University of Wisconsin, 1210 West Dayton St., Madison, WI 53706